requirements.txt
aio-pika==9.0.5
aiofiles==23.1.0
aiormq==6.7.4
dnspython==2.3.0
grpcio==1.53.0
grpcio-tools==1.53.0
html5tagger==1.3.0
httptools==0.5.0
idna==3.4
install==1.3.5
multidict==6.0.4
pamqp==3.2.1
pika==1.3.1
pymongo==4.3.3
PyYAML==6.0
sanic==23.3.0
sanic-ext==23.3.0
sanic-routing==22.8.0
tracerite==1.1.0
websockets==11.0
yarl==1.8.2
